% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{garch_settings}
\alias{garch_settings}
\title{function: GARCH Models settings.}
\usage{
garch_settings(train.size = 500, refit.every = 50, specs = list())
}
\arguments{
\item{train.size}{The training window size.}

\item{refit.every}{Determines every how many periods the garch models are re-estimated.}

\item{specs}{A named list (stock ticker => \code{\link[rugarch:uGARCHspec-class]{rugarch::uGARCHspec}} class).
If some of the tickers are left empty then a default spec will be assigned to that stock (see \link{garch_vine_roll})}
}
\description{
Method for creating a GarchSettings object.
}
