% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{garch_vine_roll}
\alias{garch_vine_roll}
\title{function: GARCH Vine Rolling Value-at-Risk Forecast and Backtesting.}
\usage{
garch_vine_roll(
  data = NULL,
  garch.settings = NULL,
  vine.settings = NULL,
  alpha = c(0.01, 0.05),
  weights = NULL
)
}
\arguments{
\item{data}{A matrix or data.frame where each column corresponds to the sorted log return series of a single stock. The row names are idially dates.}

\item{garch.settings}{A \code{\linkS4class{GarchSettings}} object specifying the settings for the ARMA-GARCH models.}

\item{vine.settings}{A \code{\linkS4class{VineSettings}} object specifying the settings for the vine copula model.}

\item{alpha}{The Value-at-Risk level to calculate.}

\item{weights}{A numerical vector specifying the weights of the portfolio components. If ommitted an equally weighted portfolio is assumed.}
}
\description{
Method for rolling forecast for the Value-at-Risk of a portfolio of assets.
}
