% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_returns}
\alias{sample_returns}
\title{A sample log returns for a few assets.}
\format{
data.frame with 3 columns and 1000 rows, where the row names represent the dates.
}
\source{
\href{https://finance.yahoo.com}{Yahoo Finance}
}
\usage{
data(sample_returns)
}
\description{
Data extracted from Yahoo Finance representing the log returns for
Google, Apple, and Amazon stocks between 2014-01-13 and 2018-01-01 where in total
there are 1000 data points.
}
\examples{
data(sample_returns)
head(sample_returns)
}
\keyword{datasets}
