% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{VaR_plot}
\alias{VaR_plot}
\title{function: Value-at-Risk plotting function.}
\usage{
VaR_plot(garch.vine.roll = NULL, alpha = 0.05)
}
\arguments{
\item{garch.vine.roll}{A \code{\linkS4class{GarchVineRoll}} object, the output of the method \code{\link{garch_vine_roll}}.}

\item{alpha}{The Value-at-Risk level.}
}
\description{
Method for plotting the Value-at-Risk forecast.
}
\examples{
library(rugarch)
library(VaRVine)
data("sample_returns")
garch.settings <- garch_settings(train.size = 750, refit.every = 50)
vine.setttings <- vine_settings(train.size = 250, refit.every = 25, family.set = 'all')
roll <- garch_vine_roll(sample_returns, garch.settings,
                     vine.settings, alpha=0.05, weights=c(0.5, 0.2, 0.3))
VaR_plot(roll, alpha = 0.05)
}
