% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vine_settings}
\alias{vine_settings}
\title{function: vine models settings.}
\usage{
vine_settings(train.size = 250, refit.every = 25, family.set = "all")
}
\arguments{
\item{train.size}{The training window size.}

\item{refit.every}{Determines every how many periods the garch models are re-estimated.}

\item{family.set}{vector of bivariate copula families. Mimics the `family_set` argument of method \code{\link[rvinecopulib:vinecop]{rvinecopulib::vinecop}}.}
}
\description{
Method for creating a VineSettings object.
}
