% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{VineSettings}
\alias{VineSettings}
\title{class: Vine Settings Class.}
\description{
Class for the vine model settings.

Class for GARCH vine rolling forecast.
}
\section{Slots}{

\describe{
\item{\code{train.size}}{The training window size.}

\item{\code{refit.every}}{Determines every how many periods the garch models are re-estimated.}

\item{\code{family.set}}{vector of bivariate copula families. Mimics the `family_set` argument of method \code{\link[rvinecopulib:vinecop]{rvinecopulib::vinecop}}.}

\item{\code{train.size}}{The training window size.}

\item{\code{refit.every}}{Determines every how many periods the garch models are re-estimated.}

\item{\code{family.set}}{vector of bivariate copula families. Mimics the `family_set` argument of method \code{\link[rvinecopulib:vinecop]{rvinecopulib::vinecop}}.}
}}

